requirements.txt
ftfy
regex
tqdm
pytorch_lightning==1.4.0
torchmetrics==0.6.0
transformers==4.8.1
Pillow==8.3.1
tqdm==4.53.0
ipdb==0.13.7
numpy==1.19.2
einops==0.3.0
pyarrow==2.0.0
sacred==0.8.2
pandas==1.1.5
fairscale==0.4.0
git+https://github.com/rwightman/pytorch-image-models.git
scipy
tensorboardX
opencv-python
datasets
nltk
git+https://github.com/microsoft/tutel@main
git+https://github.com/KaiyangZhou/Dassl.pytorch.git
vision_datasets
gdown